+++ contracts/interface/IGlobals.sol
@@ -46,5 +46,5 @@ interface IGlobals {
 
     function liquidityPoolFactory() external view returns (address);
 
-    function getPrice(address) external view returns (uint);
+    function getPrice(address) external view returns (uint256);
 }
